RFC9110 compliant exists, better setup instructions
The exists method used the reason message to determine existence on request error.
The reason message is not present in HTTP/2 and a lot of frameworks have dropped it entirely.
To handle this we now use a method of probing the endpoint to determine if an error should
be treated as indicating non-existence or not.